Disable save in progress for now
