+++ src/js/hca/config/form.js
@@ -114,8 +114,9 @@ const formConfig = {
   trackingPrefix: 'hca-',
   formId: 'hca',
   version: 0,
-  // disableSave: true,
+  disableSave: true,
   transformForSubmit: transform,
+  // TODO: When save in progress is released, change the intro page
   introduction: IntroductionPage,
   confirmation: ConfirmationPage,
   errorMessage: ErrorMessage,
